Update merge.go
Make "OverwriteWithEmptyValue" in config struct to be set by any method, so that the merge can be customizable. In cases, where request toggles a boolean value and the the request message can change a "true" value to false we can pass the flag to merge accordingly, which is not possible otherwise.