Make "OverwriteWithEmptyValue" in config struct to be set by any method. (#133)
Make "OverwriteWithEmptyValue" in config struct to be set by any method.

this make the merge more customizable. In cases, where request toggles a boolean value and the the request message can change a "true" value to false we can pass the flag to merge accordingly, which is not possible otherwise.