feat(map): Handle json.Number when merging maps
If you unmarshall a json object using the json.Decoder and the UseNumber
function, it will create numbers on the map as a `json.Number` object.
This object has helper functions for handling ints and floats. So if a
json.Number is encountered, try to handle the type of the number based
on the destination.